personal_notepad/requirements.txt
django
dj-rest-auth
django-rest-framework
django-allauth
django_resized
pillow